Add files via upload
In this project, we performed modeling of the historic sale price of the Toyota Corolla sedan and built a prediction model. To improve the prediction performance and cope with the multicollinearity issue, we utilized a ridge model, LASSO model, and variable selection method. Through the analysis, the selected ordinary least square model is found to be inadequate with deficiencies such as non-constant variance and nonlinearity. After transformation, the resulting linear model fulfills the goal and
performs much better in testing data, compared to the ridge model and the original model.
The original data comes here. https://www.kaggle.com/datasets/tolgahancepel/toyota-corolla